# Conflicts:
#	CherryCrisis/CherryEditor/imgui.ini
#	CherryCrisis/CherryEditor/src/panels/inspector.cpp
#	CherryCrisis/CherryEditor/src/panels/project_settings_displayer.cpp
#	CherryCrisis/CherryEngine/CherryEngine.vcxproj
#	CherryCrisis/CherryEngine/CherryEngine.vcxproj.filters
#	CherryCrisis/CherryEngine/include/resource_manager.inl
#	CherryCrisis/CherryEngine/include/scene.hpp
#	CherryCrisis/CherryEngine/src/model.cpp
#	CherryCrisis/CherryEngine/src/model_renderer.cpp
#	CherryCrisis/CherryEngine/src/scene.cpp
#	CherryCrisis/CherryEngine/src/transform.cpp

+++ CherryCrisis/CherryEditor/src/core/main.cpp
@@ -112,7 +112,7 @@ int main()
     while (glfwWindowShouldClose(window) == false)
     {
         InputManager::GetInstance()->UpdateKeys();
-        TimeManager::GetInstance()->Update(glfwGetTime());
+        TimeManager::GetInstance()->Update((float)glfwGetTime());
         glfwPollEvents();
 
         ImGui_ImplOpenGL3_NewFrame();
